Move box handling to world.cpp
